FLAS-44 Add file name generator move file manager
