+++ src/File/FileManagerInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Utils;
+namespace App\File;
 
 interface FileManagerInterface
 {
