+++ src/File/FileNameGeneratorInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\File;
+
+interface FileNameGeneratorInterface
+{
+    public function generate(string $identifier, string $originalFilename): string;
+}
